+++ Cinema.SQL/dbo/Tables/Movies.sql
@@ -5,6 +5,6 @@
     [Poster]        VARBINARY (MAX) NULL,
     [StartDate]     DATE            NOT NULL,
     [EndDate]       DATE            NOT NULL,
-    [MinutesLength] SMALLINT        NOT NULL,
+    [MinutesLength] INT             NOT NULL,
     CONSTRAINT [PK_Movies] PRIMARY KEY CLUSTERED ([Id] ASC)
 );
